requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.12.5
starlette==1.0.0

SQLAlchemy==2.0.49
alembic==1.18.4

python-dotenv==1.2.2

httpx==0.28.1
requests==2.33.1

passlib==1.7.4
bcrypt==5.0.0
python-jose==3.5.0

typing_extensions==4.15.0